Introduce Player module with object and object's test
The player module is required to control how two players object (which
is the maximum amount that should be created) is created and how they
will interact with each other.

	- The substantiate function will act as the constructor as a
	factory function
	- One of the player (or maybe two) can act as a bot. Both
	players object are given an attack function which could take a
	coordinate, or will automatically come up with one if it is a
	bot

Both will need to decide each other turn at the start in order to feed
into the code later on to make sure that the game itself is having
current turns.